execution/stagedsync: markDirty takes the account pointer; bench small accumulations; merge the dirty tests
